updated the make common dataset to store save the data for test/train separately

+++ make_common_dataset.py
@@ -12,11 +12,18 @@ subjects = 'all'
 signal_type = 'grad'
 n_parcels_max = 3
 no_parcels = 450
-data_dir_base = 'data/test/'
-subject_id_start = 1  # assign a number from which first subject_id will start
+data_dir_base = 'data/train/'
+# assign a number from which first subject_id will start
+if 'train' in data_dir_base:
+    subject_id_start = 1
+    subdir = 'train'
+elif 'test' in data_dir_base:
+    subject_id_start = 6
+    subdir = 'test'
 divide_to = 2  # if 1 data will not be split. Otherwise it might be 2 or 4.
 # the data will be equally split to private and public (if set 2) or
 # to private/public/test/train (if set to 4)
+data_dir_save = 'data/ramp_challenge'
 
 assert divide_to == 1 or divide_to == 2 or divide_to == 4
 
@@ -27,6 +34,8 @@ data_type = '_' + str(no_parcels) + '_' + str(n_parcels_max)
 data_dir_all = os.path.join(data_dir_base,
                             f'data_{signal_type}_all{data_type}')
 data_dir = os.path.join(data_dir_base, f'data_{signal_type}_*{data_type}')
+if data_dir_save is None:
+    data_dir_save = data_dir_all
 
 data_dirs = sorted(glob.glob(data_dir))
 # if the all_data directory is still in the list of dirs, remove it
@@ -37,18 +46,26 @@ random.seed(42)
 random.shuffle(data_dirs)
 
 
-def save_the_data(data_dir_all, data_dirs, subj_id_init):
-    # data_dir_all: where to save the data
-    # data_dirs: paths of the saved data
+def save_the_data(data_dir_all, subdir, data_dirs, subj_id_init,
+                  remove_previous=False):
+    # data_dir_all : where to save the data
+    # data_dirs : paths of the saved data
+    # remove_previous : clean up the data_dir_all directory before saving new
+    #   files
     sbj_id = 0
     # initialize the files in the data_dir_all
-    all_X_file = os.path.join(data_dir_all, 'X.csv.gz')
+    data_dir_all_subdir = os.path.join(data_dir_all, subdir)
+    all_X_file = os.path.join(data_dir_all_subdir, 'X.csv.gz')
 
     # remove previous save directory and create a clean one
-    if os.path.isdir(data_dir_all):
+    if os.path.isdir(data_dir_all) and remove_previous:
         print('removing contents from ' + data_dir_all)
         shutil.rmtree(data_dir_all)
-    os.makedirs(data_dir_all)
+    if not os.path.exists(data_dir_all):
+        os.makedirs(data_dir_all)
+
+    if not os.path.exists(data_dir_all_subdir):
+        os.makedirs(data_dir_all_subdir)
 
     for idx, subject_path in enumerate(data_dirs):
         # check if all the necessary files are present
@@ -104,9 +121,9 @@ def save_the_data(data_dir_all, data_dirs, subj_id_init):
         sbj_id += 1
     if len(data_dirs):
         # save the target
-        save_npz(os.path.join(data_dir_all, 'target.npz'), target_all)
-        print('{} samples from {} subjects were saved in the {}'.format(
-            target_all.shape[0], sbj_id, data_dir_all))
+        save_npz(os.path.join(data_dir_all_subdir, 'target.npz'), target_all)
+        print(f'{target_all.shape[0]} samples from {sbj_id} subjects'
+              f' were saved in the {data_dir_all} + {subdir}')
     else:
         print(f'Did not find any files in {data_dir}')
 
@@ -120,33 +137,39 @@ if divide_to == 4:
     # private/train equally
     n_files = int(len(data_dirs) / 4)
     # public/train
-    public_train_dir = os.path.join(data_dir_all, 'public', 'train')
-    save_the_data(public_train_dir, data_dirs[:n_files],
+    public_train_dir = os.path.join(data_dir_save, 'public')
+    save_the_data(public_train_dir, subdir='train',
+                  data_dirs=data_dirs[:n_files],
                   subj_id_init=subject_id_start)
     # public/test
-    public_test_dir = os.path.join(data_dir_all, 'public', 'test')
-    save_the_data(public_test_dir, data_dirs[n_files:n_files*2],
+    public_test_dir = os.path.join(data_dir_save, 'public')
+    save_the_data(public_test_dir, subdir='test',
+                  data_dirs=data_dirs[n_files:n_files*2],
                   subj_id_init=subject_id_start+n_files)
     # private/train
-    private_train_dir = os.path.join(data_dir_all, 'private', 'train')
-    save_the_data(private_train_dir, data_dirs[n_files*2:n_files*3],
+    private_train_dir = os.path.join(data_dir_save, 'private')
+    save_the_data(private_train_dir, subdir='train',
+                  data_dirs=data_dirs[n_files*2:n_files*3],
                   subj_id_init=subject_id_start+n_files*2)
     # private/test
-    private_test_dir = os.path.join(data_dir_all, 'private', 'test')
-    save_the_data(private_test_dir, data_dirs[n_files*3:],
+    private_test_dir = os.path.join(data_dir_save, 'private')
+    save_the_data(private_test_dir, subdir='test',
+                  data_dirs=data_dirs[n_files*3:],
                   subj_id_init=subject_id_start+n_files*3)
 elif divide_to == 2:
     # divide all the data to public and private equally
     n_files = int(len(data_dirs) / 2)
     # public/train
-    public_train_dir = os.path.join(data_dir_all, 'public')
-    save_the_data(public_train_dir, data_dirs[:n_files],
+    public_train_dir = os.path.join(data_dir_save, 'public')
+    save_the_data(public_train_dir, subdir,
+                  data_dirs=data_dirs[:n_files],
                   subj_id_init=subject_id_start)
     # public/test
-    public_test_dir = os.path.join(data_dir_all, 'private')
-    save_the_data(public_test_dir, data_dirs[n_files:],
-                  subj_id_init=subject_id_start+n_files)
+    public_test_dir = os.path.join(data_dir_save, 'private')
+    save_the_data(public_test_dir, subdir,
+                  data_dirs=data_dirs[n_files:],
+                  subj_id_init=subject_id_start)
 
 
 elif divide_to == 1:
-    save_the_data(data_dir_all, data_dirs, subj_id_init=subject_id_start)
+    save_the_data(data_dir_save, data_dirs, subj_id_init=subject_id_start)
